Few improvements

+++ src/main/scala/org/github/insider/alchemy/repository/TradesRepository.scala
@@ -3,5 +3,5 @@ package org.github.insider.alchemy.repository
 import org.github.insider.polymarket.domain.Trade
 
 trait TradesRepository[F[_]] {
-  def insert(trades: List[Trade]): F[Long]
+  def insert(trades: List[Trade]): F[Int]
 }
